refactor: widgets/sidbar 컴포넌트들 ui 폴더로 이동
