+++ client/src/widgets/sidebar/index.ts
@@ -1,1 +1,1 @@
-export { Sidebar } from './Sidebar';
+export { Sidebar } from './ui/Sidebar';
