+++ client/src/widgets/sidebar/ui/Credit.tsx
