+++ client/src/widgets/sidebar/ui/StreamingList.css
