+++ client/src/widgets/sidebar/ui/StreamingList.tsx
