+++ client/src/widgets/sidebar/ui/StreamingListItem.tsx
